chore(query): make soundex a mod
